fix: complete and test event creation workflow
- Fixed organizer foreign key and DTO mapping issues
- Resolved POST request 401/500 errors
- Initialized ticket and status defaults correctly
- Verified API behavior via Postman testing

+++ src/main/java/com/smart/event/event_management_system/service/BookingService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.smart.event.event_management_system.entity.Booking;
 
 public interface BookingService {
-    Booking createBooking(Long userId, Integer eventId, int tickets);
+    Booking createBooking(Long userId, Long eventId, int tickets);
     List<Booking> getBookingsByUser(Long userId);
     List<Booking> getBookingsByEvent(Integer eventId);
     void cancelBooking(Integer bookingId);
